FileUpload component
Factor out FileUpload to its own component.  Include the declared type of uploaded objects in the app-specific header of `x-sos-content-type`
